requirements.txt
Django==3.1.2

black==20.8b1
click==7.1.2
mypy-extensions==0.4.3
pathspec==0.8.0
regex==2020.10.28
typed-ast==1.4.1
typing-extensions==3.7.4.3

isort==5.6.4

pre-commit==2.8.2
appdirs==1.4.4
cfgv==3.2.0
distlib==0.3.1
filelock==3.0.12
identify==1.5.9
nodeenv==1.5.0
pyyaml==5.3.1
six==1.15.0
toml==0.10.2
virtualenv==20.1.0

pytest==6.1.2
attrs==20.3.0
iniconfig==1.1.1
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7

pytest-django==4.1.0

django-extensions==3.0.9

djangorestframework==3.12.2

django-filter==2.4.0

drf-yasg==1.20.0
MarkupSafe==1.1.1
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
jinja2==2.11.2
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
uritemplate==3.0.1
urllib3==1.25.11

parameterized==0.7.4
